moving deprecation decorator to uitl.py
